requirements.txt
asgiref==3.8.1
Django==4.2.19
django-extensions==3.2.3
djangorestframework==3.15.2
python-environ==0.4.54
sqlparse==0.5.3
typing_extensions==4.12.2
